feat: RESKIN Cookies - Update SkyButton component
